backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]>=0.31.0

# Data processing
pandas==2.2.0
openpyxl==3.1.2
numpy==1.26.0

# Validation and settings
pydantic>=2.7.0
pydantic-settings==2.5.2
python-dotenv==1.0.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart>=0.0.9

# Email
aiosmtplib==3.0.1
email-validator==2.1.0

# MCP (Model Context Protocol)
mcp==1.20.0

# Database (optional)
sqlalchemy==2.0.25
alembic==1.13.1

# Testing
pytest==7.4.3
pytest-asyncio==0.23.3
httpx>=0.27.1

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# OTP
pyotp==2.9.0
